Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Martin/Alex/Damian,I will have Santanu explain the differences. Obviously, because it is Postman, he does not have control over what it might generate as diffs.
> 
> To the point of using variables, the idea of providing these scripts was meant to show how a model could be validated. It was never mean to *be* the way to validate the Fronthaul interface.

Well, this has little to do with variables actually.
If that collection shall serve as a reference for anything, then it would be nice to make sure, that it could be re-used in different environment. Variables are the way to make that easier.

If the intention is to demonstrate a real-world example, then the question would be - shall this really be a code in a repository? Maybe a Confluence page is better, especially, if You don`t intendt to modify this collection later on...?

Patch-set: 1
Reviewer: Gerrit User 1001144 <1001144@84e4e190-e623-4c37-9eda-58b1937c1b53>